requirements.txt
Django==4.2.16
psycopg2-binary==2.9.9
python-decouple==3.8
stripe==7.8.0
Pillow==10.2.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-extensions==3.2.3
whitenoise==6.6.0

